# swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Motozalozna backend API",
    "version" : "0.0.1",
    "title" : "Motozalozna API",
    "contact" : {
      "email" : "juhas.jugi@gmail.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "motozalozna-backend.herokuapp.com",
  "basePath" : "/api",
...
  "schemes" : [ "http" ],
  "paths" : {
...
    "/form/check_price" : {
      "post" : {
        "tags" : [ "form" ],
        "summary" : "Returs price of car",
        "operationId" : "checkPrice",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "form-data",
          "description" : "All data from car form.",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Price_Check"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "return price"
          },
          "500" : {
            "description" : "server error"
          }
        }
      }
    },
    "/form/check_stolen" : {
      "post" : {
        "tags" : [ "form" ],
        "summary" : "Returs if car is stolen",
        "operationId" : "checkStolen",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "ecv",
          "description" : "ECV of car.",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Stolen_Check"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "return number value"
          },
          "500" : {
            "description" : "server error"
          }
        }
      }
    }
  },
  "definitions" : {
...
    "Price_Check" : {
      "properties" : {
        "karoseria" : {
          "type" : "number",
          "example" : 1.0
        },
        "palivo" : {
          "type" : "number",
          "example" : 1.0
        },
        "pohon" : {
          "type" : "number",
          "example" : 1.0
        },
        "prevodovka" : {
          "type" : "number",
          "example" : 0
        },
        "vykon" : {
          "type" : "number",
          "example" : 110
        },
        "vek" : {
          "type" : "number",
          "example" : 3.0
        },
        "pocetkm" : {
          "type" : "number",
          "example" : 20000
        },
        "auto" : {
          "type" : "number",
          "example" : 168
        }
      }
    },
    "Stolen_Check" : {
      "properties" : {
        "ecv" : {
          "type" : "string",
          "example" : "VT657CG"
        }
      }
    }
  }
}